Follow recommended pyadjoint interface Relocate UFL Constrainst to dolfin_common
