+++ firedrake_adjoint/__init__.py
@@ -19,7 +19,9 @@ from pyadjoint.drivers import compute_gradient, compute_hessian
 from pyadjoint.adjfloat import AdjFloat
 from pyadjoint.control import Control
 from pyadjoint import IPOPTSolver, ROLSolver, MinimizationProblem, InequalityConstraint, minimize
-from pyadjoint.ufl_constraints import UFLInequalityConstraint, UFLEqualityConstraint
+
+#from pyadjoint.ufl_constraints import UFLInequalityConstraint, UFLEqualityConstraint
+from dolfin_adjoint_common.ufl_constraints import UFLInequalityConstraint, UFLEqualityConstraint
 import numpy_adjoint
 
 continue_annotation()
